Fix error due to large properties set
When the list of properties is too large the API raises <FatalAPI ERROR
400>.
To fix this issue some of the methods in the main HubspotStream
class were overwritten to cover such cases by returning a generator
instead of the whole list. This way all the properties are passed in
batches.